Remove old veda-data utils file and consolidate all methods to new directory independent of faux modules
