+++ app/scripts/components/common/dataset-classification.tsx
@@ -13,11 +13,17 @@ import { CollecticonMedal } from './icons/medal';
 import { DatasetData } from '$types/veda';
 
 import { variableGlsp } from '$styles/variable-utils';
+// import {
+//   getTaxonomy,
+//   TAXONOMY_GRADE,
+//   TAXONOMY_UNCERTAINTY
+// } from '$utils/veda-data-no-faux-module';
+
 import {
   getTaxonomy,
   TAXONOMY_GRADE,
   TAXONOMY_UNCERTAINTY
-} from '$utils/veda-data-no-faux-module';
+} from '$utils/veda-data/taxonomies';
 
 const DATA_UNCERTAINTY = {
   High: CollecticonProgressTickLow,
