+++ app/scripts/components/common/mdx-content.js
@@ -2,7 +2,8 @@ import React from 'react';
 import T from 'prop-types';
 import { MDXProvider } from '@mdx-js/react';
 
-import { useMdxPageLoader } from '$utils/veda-data';
+// import { useMdxPageLoader } from '$utils/veda-data';
+import { useMdxPageLoader } from '$utils/veda-data/mdx';
 import { S_LOADING, S_SUCCEEDED } from '$utils/status';
 import { ContentLoading } from '$components/common/loading-skeleton';
 import Block from '$components/common/blocks';
