+++ app/scripts/components/exploration/data-utils.ts
@@ -19,14 +19,14 @@ import {
   DATA_METRICS,
   DEFAULT_DATA_METRICS
 } from './components/datasets/analysis-metrics';
-import { veda_datasets } from '$data-layer/datasets';
+import { veda_faux_module_datasets } from '$data-layer/datasets';
 
 import { utcString2userTzDate } from '$utils/date';
 
-// @TODO: These should be updated to take in datasets as a param instead of using veda_datasets directly
+// @TODO: These should be updated to take in datasets as a param instead of using veda_faux_module_datasets directly
 
 export const findParentDataset = (layerId: string) => {
-  const parentDataset: VedaDatum<DatasetData> | undefined = Object.values(veda_datasets).find((dataset: VedaDatum<DatasetData>) =>
+  const parentDataset: VedaDatum<DatasetData> | undefined = Object.values(veda_faux_module_datasets).find((dataset: VedaDatum<DatasetData>) =>
     dataset!.data.layers.find((l) => l.id === layerId)
   );
   return parentDataset?.data;
@@ -39,14 +39,15 @@ export const findDatasetAttribute = ({
   datasetId: string;
   attr: string;
 }) => {
-  return veda_datasets[datasetId]?.data[attr];
+  return veda_faux_module_datasets[datasetId]?.data[attr];
 };
 
-export const allDatasets = Object.values(veda_datasets)
+// @NOTE: This exists in "app/scripts/utils/veda-data-no-faux-module", move all references to use that instead
+export const allDatasets = Object.values(veda_faux_module_datasets)
   .map((d: VedaDatum<DatasetData>) => d!.data);
 
 
-export const allExploreDatasets = Object.values(veda_datasets)
+export const allExploreDatasets = Object.values(veda_faux_module_datasets)
   .map((d: VedaDatum<DatasetData>) => d!.data)
   .filter((d: DatasetData) => !d.disableExplore);
 
@@ -74,7 +75,7 @@ export const getAllDatasetsWithEnhancedLayers = (
   dataset
 ): DatasetDataWithEnhancedLayers[] => dataset.map(enhanceDatasetLayers);
 
-export const datasetLayers = Object.values(veda_datasets)
+export const datasetLayers = Object.values(veda_faux_module_datasets)
   .flatMap((dataset: VedaDatum<DatasetData>) => {
     return dataset!.data.layers.map(l => ({
       ...l,
