+++ app/scripts/components/stories/single/index.tsx
@@ -6,13 +6,16 @@ import PageHero from '$components/common/page-hero';
 import { PageMainContent } from '$styles/page';
 import RelatedContent from '$components/common/related-content';
 import { STORIES_PATH } from '$utils/routes';
+// import { useStory } from '$utils/veda-data-to-deprecate';
 import { useStory } from '$utils/veda-data';
 import { ContentTaxonomy } from '$components/common/content-taxonomy';
+import { veda_faux_module_stories } from '$data-layer/datasets';
 
 const MdxContent = lazy(() => import('$components/common/mdx-content'));
 
 function StoriesSingle() {
-  const story = useStory();
+  // const story = useStory();
+  const story = useStory(veda_faux_module_stories);
 
   if (!story || story.data.asLink) throw resourceNotFound();
